+++ core/src/main/java/Gameplay/Movimientos/MovimientoRango.java
@@ -6,18 +6,15 @@ import entidades.personajes.Personaje;
 
 public abstract class MovimientoRango extends Movimiento {
     protected float velocidad;
-    protected int danio;
     protected GestorProyectiles gestorProyectiles;
 
-    public MovimientoRango(String nombre, Texture textura, int cooldown, float velocidad, int danio, GestorProyectiles gestorProyectiles) {
+    public MovimientoRango(String nombre, Texture textura, int cooldown, float velocidad, GestorProyectiles gestorProyectiles) {
         super(nombre, textura, cooldown);
         this.velocidad = velocidad;
-        this.danio = danio;
         this.gestorProyectiles = gestorProyectiles;
     }
 
     public abstract void ejecutar(Personaje personaje, float potencia);
 
     public float getVelocidad() { return velocidad; }
-    public int getDano() { return danio; }
 }